Introduce abstract id class
